+++ python/mlcroissant/mlcroissant/__init__.py
@@ -34,5 +34,4 @@ __all__ = [
     "Source",
     "Transform",
     "ValidationError",
-    "FileObject",
 ]
