+++ wizard/utils.py
@@ -1,4 +1,8 @@
 DF_HEIGHT = 150
+from core.state import CurrentStep
+from core.state import Metadata
+import streamlit as st
+
 import mlcroissant as mlc
 
 EDITOR_CACHE = mlc.constants.CROISSANT_CACHE / "editor"
@@ -7,3 +11,19 @@ LOADED_CROISSANT = EDITOR_CACHE / "loaded_croissant"
 
 def needed_field(text: str) -> str:
     return f"{text}:red[*]"
+
+def set_form_step(action, step=None):
+    """Maintains the user's location within the wizard."""
+    if action == "Jump" and step is not None:
+        st.session_state[CurrentStep] = step
+
+def init_state():
+
+    if Metadata not in st.session_state:
+        st.session_state[Metadata] = Metadata()
+
+    if mlc.Dataset not in st.session_state:
+        st.session_state[mlc.Dataset] = None
+
+    if CurrentStep not in st.session_state:
+        st.session_state[CurrentStep] = CurrentStep.start
