+++ wizard/views/record_sets.py
@@ -16,6 +16,7 @@ def render_record_sets():
         st.markdown("Please add files first.")
     else:
         for record_set in st.session_state[Metadata].record_sets:
+            st.warning(record_set)
             record_set_conv = pd.DataFrame(record_set.__dict__)
             with st.container():
